Update: ProductCell에 행사 종료 레이어 추가
