+++ Pyonsnal-Color/Pyonsnal-Color/Common/Cell/ProductCell/ProductCell+ViewHolder.swift
@@ -1,33 +1,15 @@
 //
-//  ProductCell.swift
+//  ProductCell+ViewHolder.swift
 //  Pyonsnal-Color
 //
-//  Created by 조소정 on 2023/06/14.
+//  Created by 조소정 on 2023/09/19.
 //
 
 import UIKit
-import Combine
-import SnapKit
 
-protocol ProductCellDelegate: AnyObject {
-    func didTapFavoriteButton(productId: String, action: FavoriteButtonAction)
-}
-
-final class ProductCell: UICollectionViewCell {
-    
+extension ProductCell {
     enum Size {
-        // TO DO : component margin 정리 주시면 CommonMarginManager로 분리 예정
-        static let dividerMargin: CGFloat = 12
         static let productImageViewMargin: CGFloat = 25.5
-        static let convenientTagImageViewMargin: CGFloat = 12
-        static let tagImageViewMargin: CGFloat = 12
-        static let newImageViewMargin: CGFloat = 12
-        static let titleLabelLeading: CGFloat = 4
-        static let titleLabelMargin: CGFloat = 12
-        static let priceContainerViewTop: CGFloat = 4
-        static let priceContainerViewMargin: CGFloat = 12
-        static let discountPriceLabelLeading: CGFloat = 4
-        
         static let dividerHeight: CGFloat = 1
         static let productImageContainerViewHeight: CGFloat = 160
         static let convenientTagImageViewWidth: CGFloat = 36
@@ -40,13 +22,6 @@ final class ProductCell: UICollectionViewCell {
         static let borderWidth: CGFloat = 1
         static let favoriteButtonSize: CGFloat = 20
     }
-    
-    // MARK: - Private Property
-    private let viewHolder: ViewHolder = .init()
-    private var cancellable = Set<AnyCancellable>()
-    private var productId: String?
-    
-    weak var delegate: ProductCellDelegate?
     
     final class ViewHolder: ViewHolderable {
         
@@ -141,6 +116,22 @@ final class ProductCell: UICollectionViewCell {
             return button
         }()
         
+        let eventCloseLayerView: UIView = {
+            let view = UIView()
+            view.backgroundColor = .black.withAlphaComponent(0.2)
+            view.isHidden = true
+            view.layer.masksToBounds = true
+            return view
+        }()
+        
+        let eventCloseLabel: UILabel = {
+            let label = UILabel()
+            label.text = "행사 종료"
+            label.textColor = .white
+            label.font = .title2
+            return label
+        }()
+        
         // MARK: - Method
         func place(in view: UIView) {
             view.addSubview(stackView)
@@ -162,6 +153,10 @@ final class ProductCell: UICollectionViewCell {
             
             priceContainerView.addSubview(originalPriceLabel)
             priceContainerView.addSubview(discountPriceLabel)
+            
+            // close Layer
+            view.addSubview(eventCloseLayerView)
+            eventCloseLayerView.addSubview(eventCloseLabel)
         }
         
         func configureConstraints(for view: UIView) {
@@ -185,7 +180,7 @@ final class ProductCell: UICollectionViewCell {
             }
             
             convenienceStoreTagImageView.snp.makeConstraints {
-                $0.top.leading.equalToSuperview().inset(Size.convenientTagImageViewMargin)
+                $0.top.leading.equalToSuperview().inset(.spacing12)
                 $0.width.height.equalTo(Size.convenientTagImageViewWidth)
             }
             
@@ -195,10 +190,7 @@ final class ProductCell: UICollectionViewCell {
                 $0.height.equalTo(Size.eventTagImageViewHeight)
             }
             
-            favoriteButton.snp.makeConstraints {
-                $0.size.equalTo(Size.favoriteButtonSize)
-                $0.top.trailing.equalToSuperview().inset(.spacing12)
-            }
+            self.updateFavoriteButtonConstraints()
             
             itemInfoContainerView.snp.makeConstraints {
                 $0.leading.equalTo(stackView)
@@ -216,12 +208,12 @@ final class ProductCell: UICollectionViewCell {
             
             titleLabel.snp.makeConstraints {
                 $0.height.equalTo(titleLabel.font.customLineHeight)
-                $0.leading.equalTo(newTagView.snp.trailing).offset(Size.titleLabelLeading)
+                $0.leading.equalTo(newTagView.snp.trailing).offset(.spacing4)
             }
             
             priceContainerView.snp.makeConstraints {
-                $0.top.equalTo(itemInfoTopStackView.snp.bottom).offset(Size.priceContainerViewTop)
-                $0.leading.trailing.bottom.equalToSuperview().inset(Size.priceContainerViewMargin)
+                $0.top.equalTo(itemInfoTopStackView.snp.bottom).offset(.spacing4)
+                $0.leading.trailing.bottom.equalToSuperview().inset(.spacing12)
             }
             
             originalPriceLabel.snp.contentHuggingHorizontalPriority = 251
@@ -232,97 +224,25 @@ final class ProductCell: UICollectionViewCell {
             }
             
             discountPriceLabel.snp.makeConstraints {
-                $0.leading.equalTo(originalPriceLabel.snp.trailing).offset(Size.discountPriceLabelLeading)
+                $0.leading.equalTo(originalPriceLabel.snp.trailing).offset(.spacing4)
                 $0.top.trailing.bottom.equalToSuperview()
             }
-        }
-    }
-    
-    // MARK: - Initializer
-    override init(frame: CGRect) {
-        super.init(frame: .zero)
-        configureUI()
-        viewHolder.place(in: contentView)
-        viewHolder.configureConstraints(for: contentView)
-        bindActions()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    func bindActions() {
-        viewHolder.favoriteButton
-            .tapPublisher
-            .throttle(for: 0.5, scheduler: RunLoop.main, latest: false)
-            .sink { [weak self] _ in
-                guard let self, let productId else { return }
-                let isSelected = !viewHolder.favoriteButton.isSelected
-                self.setFavoriteButtonSelected(isSelected: isSelected)
-                
-                let action: FavoriteButtonAction = getFavoriteButtonSelected() ? .add : .delete
-                self.delegate?.didTapFavoriteButton(productId: productId, action: action)
-            }.store(in: &cancellable)
-    }
-    
-    func updateCell(with product: ProductConvertable?) {
-        guard let product else { return }
-        self.productId = product.productId
-        viewHolder.titleLabel.text = product.name
-        if let storeTypeImage = product.storeType.storeTagImage {
-            viewHolder.convenienceStoreTagImageView.setImage(storeTypeImage)
-        }
-        viewHolder.itemImageView.setImage(with: product.imageURL)
-        viewHolder.originalPriceLabel.text = product.price.addWon()
-        viewHolder.newTagView.isHidden = !(product.isNew ?? false)
-        if product.isNew ?? false {
-            viewHolder.newTagView.snp.updateConstraints { make in
-                make.width.equalTo(Size.newImageViewWidth)
+            
+            eventCloseLayerView.snp.makeConstraints {
+                $0.edges.equalToSuperview()
             }
             
-            viewHolder.titleLabel.snp.updateConstraints { make in
-                make.leading.equalTo(viewHolder.newTagView.snp.trailing).offset(Size.titleLabelLeading)
-            }
-        } else {
-            viewHolder.newTagView.snp.updateConstraints { make in
-                make.width.equalTo(0)
-            }
-            viewHolder.titleLabel.snp.updateConstraints { make in
-                make.leading.equalTo(viewHolder.newTagView.snp.trailing)
+            eventCloseLabel.snp.makeConstraints {
+                $0.center.equalToSuperview()
             }
         }
         
-        hasEventType(product.eventType)
-    }
-    
-    // MARK: - Private Method
-    private func configureUI() {
-        let attributedText = viewHolder.discountPriceLabel.text?.strikeThrough(with: .gray500)
-        viewHolder.discountPriceLabel.attributedText = attributedText
-        viewHolder.convenienceStoreTagImageView.makeRounded(with: Size.convenientTagImageViewWidth / 2)
-        makeRounded(with: Size.cornerRadius)
-        makeBorder(width: Size.borderWidth, color: UIColor.gray200.cgColor)
-        setFavoriteButton(isVisible: true)
-    }
-    
-    private func hasEventType(_ event: EventTag?) {
-        if let eventName = event?.name, !eventName.isEmpty {
-            viewHolder.eventTagLabel.isHidden = false
-            viewHolder.eventTagLabel.text = eventName
-        } else {
-            viewHolder.eventTagLabel.isHidden = true
+        func updateFavoriteButtonConstraints() {
+            favoriteButton.snp.makeConstraints {
+                $0.size.equalTo(Size.favoriteButtonSize)
+                $0.top.trailing.equalToSuperview().inset(.spacing12)
+            }
         }
     }
-    
-    func setFavoriteButton(isVisible: Bool) {
-        viewHolder.favoriteButton.isHidden = !isVisible
-    }
-    
-    func setFavoriteButtonSelected(isSelected: Bool) {
-        viewHolder.favoriteButton.isSelected = isSelected
-    }
-    
-    func getFavoriteButtonSelected() -> Bool {
-        return viewHolder.favoriteButton.isSelected
-    }
+
 }
